+++ CHANGELOG.md
@@ -8,6 +8,17 @@ Noteworthy Changes:
 Changes:
 
 * Address pending octoDNS 2.x deprecations, require minimum of 1.5.x
+* Correctly quote and chunk TXT records to match Cloudflare's internal behavior
+
+## v0.0.9 - 2025-02-06 - Unknown nameservers are a thing
+
+* Handle cases where Cloudflare doesn't return a zones name servers.
+
+## v0.0.8 - 2025-02-06 - More options
+
+* Add support for optionally retrying requests that hit 403 errors
+* Add a zone_id lookup fallback when deleting records
+* Add support for setting Cloudflare plan type for zones
 
 ## v0.0.7 - 2024-08-20 - DS always come second